Added part_of tests on groups.
