+++ src/niweb/apps/noclook/tests/schema/test_metatypes.py
@@ -3,6 +3,9 @@ __author__ = 'ffuentes'
 
 from apps.noclook.tests.schema.base import Neo4jGraphQLGenericTest
 from apps.noclook.tests.stressload.data_generator import *
+from collections import OrderedDict
+from graphene import relay
+from niweb.schema import schema
 
 class Neo4jGraphQLMetatypeTest(Neo4jGraphQLGenericTest):
     pass
@@ -42,7 +45,28 @@ class Neo4jGraphQLGroupTest(Neo4jGraphQLLogicalTest):
         self.assertFalse(has_relation)
 
         # on the graphql api
+        group_id = relay.Node.to_global_id(
+            'Group', str(group.handle_id))
 
+        query = """
+        {{
+          getGroupById(id: "{group_id}"){{
+            id
+            name
+            part_of{{
+              id
+              name
+            }}
+          }}
+        }}
+        """.format(group_id=group_id)
+        result = schema.execute(query, context=self.context)
+        assert not result.errors, pformat(result.errors, indent=1)
+
+        has_relation = result.data['getGroupById']['part_of'] != None
+        self.assertFalse(has_relation)
+
+        # add relation
         relation_maker.add_part_of(group, physical_node)
 
         # check that the relation exists now
@@ -54,3 +78,8 @@ class Neo4jGraphQLGroupTest(Neo4jGraphQLLogicalTest):
         self.assertTrue(has_relation)
 
         # on the graphql api
+        result = schema.execute(query, context=self.context)
+        assert not result.errors, pformat(result.errors, indent=1)
+
+        has_relation = result.data['getGroupById']['part_of'] != None
+        self.assertTrue(has_relation)
